Added editor camera.

+++ ModernEngine/src/ModernEngine/Core/MouseButtonCodes.h
@@ -22,7 +22,7 @@ namespace ModernEngine
 
 	inline std::ostream& operator<<(std::ostream& os, MouseCode mouseCode)
 	{
-		os << static_cast<int32_t>(mouseCode);
+		os << mouseCode;
 		return os;
 	}
 }
